Revert "Cleared out the old django view templates and added routing + partials for angular"
This reverts commit 9d0a9cdb02d8ccb0952f4f8732f9740fdb14a934.

I like where this was going but it broke the UI - filtering didn't work
anymore. I tried to understand what was happening but it's not clear to
me, so I just reverted it instead. Please test the UI with fake_data and
make sure it's changing when you select different filters before trying
to land this again.